+++ 2020-01-Winter/5_graphs/.vscode/settings.json
@@ -40,6 +40,10 @@
         "sstream": "cpp",
         "stdexcept": "cpp",
         "streambuf": "cpp",
-        "typeinfo": "cpp"
+        "typeinfo": "cpp",
+        "cinttypes": "cpp",
+        "cstring": "cpp",
+        "functional": "cpp",
+        "iterator": "cpp"
     }
 }
