+++ 2020-01-Winter/5_graphs/Edge.h
@@ -1,0 +1,16 @@
+#ifndef EDGE_H
+#define EDGE_H
+
+class Vertex;
+
+class Edge
+{
+    friend class Vertex;
+public:
+    Edge(Vertex *BeingCalled) : Called(BeingCalled){};
+
+private:
+    Vertex *Called;
+};
+
+#endif
